refactor: relocate ws server
