+++ Server/network/__init__.py
@@ -1,0 +1,1 @@
+"""Network package for server."""
