+++ Server/network/ws_server.py
@@ -1,18 +1,18 @@
-import asyncio
-import socket
-import json
-import time
+from __future__ import annotations
+import asyncio, json, socket, time
+from typing import Callable, Optional
 import websockets
+from ..App.services.vision_service import VisionService
 
-from core.VisionInterface import VisionInterface
-from core.vision import api as vision_api
+async def _wait_for_frame(svc: VisionService, timeout=3.0, poll=0.05) -> Optional[str]:
+    deadline = time.monotonic() + float(timeout)
+    img = svc.last_b64()
+    while img is None and time.monotonic() < deadline:
+        await asyncio.sleep(poll)
+        img = svc.last_b64()
+    return img
 
-camera = VisionInterface()
-camera.start()
-camera.start_stream(interval_sec=1.0)  # sigue autoarrancando; si prefieres lazy, quita esta línea
-
-
-def get_local_ip():
+def _get_local_ip() -> str:
     try:
         s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
         s.connect(("8.8.8.8", 80))
@@ -22,85 +22,47 @@ def get_local_ip():
     except Exception:
         return "127.0.0.1"
 
+def make_handler(svc: VisionService) -> Callable:
+    async def handler(ws):
+        print("[WS] client connected")
+        async for message in ws:
+            try:
+                data = json.loads(message)
+                cmd = data.get("cmd")
 
-async def wait_for_frame(timeout=3.0, poll=0.05):
-    """Wait until timeout for the camera to have a processed frame."""
-    deadline = time.monotonic() + float(timeout)
-    img_str = camera.get_last_processed_encoded()
-    while img_str is None and time.monotonic() < deadline:
-        await asyncio.sleep(poll)
-        img_str = camera.get_last_processed_encoded()
-    return img_str
-
-
-async def handler(websocket):
-    print("[WS] New client connected")
-    async for message in websocket:
-        try:
-            data = json.loads(message)
-            cmd = data.get("cmd")
-            if cmd == "ping":
-                response = {"status": "ok", "type": "text", "data": "pong"}
-
-            elif cmd == "start":
-                interval = float(data.get("interval", 1.0))
-                camera.start()
-                camera.start_stream(interval_sec=interval)
-                response = {"status": "ok", "type": "text", "data": f"capture started @ {interval}s"}
-
-            elif cmd == "stop":
-                camera.stop()
-                response = {"status": "ok", "type": "text", "data": "capture stopped"}
-
-            elif cmd == "capture":
-                # intenta devolver frame; si aún no hay, espera un poco
-                img_str = await wait_for_frame(timeout=float(data.get("timeout", 2.0)))
-                if img_str is None:
-                    response = {"status": "wait", "type": "text", "data": "no frame yet"}
+                if cmd == "ping":
+                    resp = {"status": "ok", "type": "text", "data": "pong"}
+                elif cmd == "start":
+                    interval = float(data.get("interval", 1.0))
+                    svc.start(interval_sec=interval)
+                    resp = {"status": "ok", "type": "text", "data": f"capture started (interval={interval}s)"}
+                elif cmd == "stop":
+                    svc.stop()
+                    resp = {"status": "ok", "type": "text", "data": "capture stopped"}
+                elif cmd == "capture":
+                    img = await _wait_for_frame(svc, float(data.get("timeout", 2.0)))
+                    resp = {"status": "ok" if img else "wait",
+                            "type": "image" if img else "text",
+                            "data": img or "no frame yet"}
+                elif cmd == "process":
+                    svc.set_processing(data)
+                    resp = {"status": "ok", "type": "text", "data": "processing updated"}
                 else:
-                    response = {"status": "ok", "type": "image", "data": img_str}
+                    resp = {"status": "error", "type": "text", "data": f"unknown command: {cmd}"}
 
-            elif cmd == "process":
-                # filtra solo claves conocidas; pasa al pipeline vía set_processing_config
-                allowed = {"blur", "edges", "contours", "ref_size"}
-                config = {k: v for k, v in data.items() if k in allowed}
-                camera.set_processing_config(config)
-                response = {"status": "ok", "type": "text", "data": "processing config updated"}
+                await ws.send(json.dumps(resp))
+            except Exception as e:
+                await ws.send(json.dumps({"status": "error", "type": "text", "data": str(e)}))
+    return handler
 
-            elif cmd == "load_profile":
-                which = data.get("which", "big")
-                path = data.get("path")
-                vision_api.load_profile(which, path)
-                response = {"status": "ok", "type": "text", "data": f"profile {which} loaded"}
-
-            elif cmd == "dynamic":
-                which = data.get("which", "big")
-                params = data.get("params", {})
-                vision_api.update_dynamic(which, params)
-                response = {"status": "ok", "type": "text", "data": "dynamic params updated"}
-
-            else:
-                response = {"status": "error", "type": "text", "data": f"unknown command: {cmd}"}
-
-            await websocket.send(json.dumps(response))
-
-        except Exception as e:
-            await websocket.send(json.dumps({"status": "error", "type": "text", "data": str(e)}))
-
-
-async def start_ws_server_async():
-    addr = get_local_ip()
-    print(f"WebSocket listening in ws://{addr}:8765 ...")
-    async with websockets.serve(handler, "0.0.0.0", 8765):
+async def start_ws_server_async(svc: VisionService, host: str = "0.0.0.0", port: int = 8765) -> None:
+    addr = _get_local_ip()
+    print(f"WebSocket listening at ws://{addr}:{port} ...")
+    async with websockets.serve(make_handler(svc), host, port):
         try:
-            await asyncio.Future()  # keep server running
-        except asyncio.CancelledError:
-            print("Server stopped with Ctrl+C.")
+            await asyncio.Future()
         finally:
-            # parada limpia de la cámara
-            camera.stop()
+            pass
 
-
-def start_ws_server():
-    asyncio.run(start_ws_server_async())
-
+def start_ws_server(svc: VisionService, host: str = "0.0.0.0", port: int = 8765) -> None:
+    asyncio.run(start_ws_server_async(svc, host=host, port=port))
